created teacher list search not working at the moment
