+++ src/frontend/src/components/teacher/index.ts
@@ -1,2 +1,1 @@
 export * from "./TeacherTable"
-export * from "./TeacherSearch"
