+++ src/frontend/src/hooks/use-query/use-teacher/index.ts
@@ -1,2 +1,3 @@
 export * from "./useList"
 export * from "./useSearch"
+export * from "./useSearchSubmit"
